Add AssetMapper with injected PHImageManager
