+++ CodingSession/GalleryViewModel.swift
@@ -41,31 +41,3 @@ class GalleryViewModel {
         self.cellControllers.accept(cellControllers)
     }
 }
-
-struct AssetMapper {
-    static func map(asset: PHAsset, imageSize: Double) -> CellController {
-        let manager = PHImageManager.default()
-        let requestOptions = PHImageRequestOptions()
-        requestOptions.isSynchronous = false
-        requestOptions.deliveryMode = .highQualityFormat
-        
-        let formatter = DateComponentsFormatter()
-        formatter.allowedUnits = [.hour, .minute, .second]
-        formatter.zeroFormattingBehavior = [.pad]
-        formatter.unitsStyle = .positional
-        
-        let title = formatter.string(from: asset.duration) ?? ""
-        
-        let controller = CellController(title: title) {
-            let resultImageData = await withCheckedContinuation { continuation in
-                manager.requestImage(for: asset, targetSize: CGSize(width: imageSize, height: imageSize), contentMode: .aspectFill, options: requestOptions) { (image, _) in
-                    continuation.resume(returning: image?.pngData())
-                }
-            }
-            
-            return resultImageData ?? Data()
-        }
-        
-        return controller
-    }
-}
